Remove useless code after USB transfer
This code was incorrect in the case where the bytes written during
tud_audio_write() was more than the side of the endpoint. That case
actually never happens because we always write less than
USB_AUDIO_PAYLOAD_SIZE and USB_AUDIO_PAYLOAD_SIZE is actually the size
of the endpoint.

Still, the code as it was has no real effect and it's error-prone in
case certain transfer parameters change.